Change data types for the WimaArReg struct
This is in preparation for handling the struct. It is also to make it so
the struct does not have 7 bytes padding.

+++ src/area.h
@@ -103,10 +103,10 @@ typedef struct WimaArReg {
 	/// The current size of the region.
 	/// This only applies to the dimension
 	/// opposite the vertical flag.
-	int size;
+	uint32_t size;
 
 	/// The min size of the region.
-	int min;
+	uint16_t min;
 
 	/// The region flags.
 	uint8_t flags;
